requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
confluent-kafka==2.3.0
google-cloud-aiplatform==1.38.0
requests==2.31.0
python-dateutil==2.8.2
elevenlabs==0.2.27
